pyproject.toml
[tool.poetry]
name = "gatekeeper_grpc"
version = "0.1.0"
description = "gRPC backend for SAI Gatekeeper service."
authors = ["Stephen Onnen <stephen.onnen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
protobuf = "^3.11.2"
grpcio = "^1.26.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
isort = "^4.3"
coverage = "^4.5"
flake8 = "^3.7.9"

"RPI.GPIO" = {version = "^0.7.0", optional=true}
rpi_ws281x = {version = "^4.2", optional=true}
adafruit-circuitpython-neopixel = {version = "^3.4", optional=true}

[tool.poetry.extras]
rpi = ["RPI.GPIO", "rpi_ws281x", "adafruit-circuitpython-neopixel"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
